refactor: menu
