+++ src/components/Sections/Navbar/Menu/MenuColumn/LinkItem/index.ts
@@ -1,0 +1,1 @@
+export { LinkItem } from './LinkItem';
