+++ src/components/Sections/Navbar/Menu/MenuColumn/LinkItem/types.ts
@@ -1,0 +1,6 @@
+import { LinkItem } from '../../types';
+
+export interface ILinkItem {
+  data: LinkItem;
+  closeOverlay: CallableFunction;
+}
