+++ src/components/Sections/Navbar/Menu/MenuColumn/RouterNavItem/ToggleList/types.ts
@@ -1,0 +1,6 @@
+import { ServiceItem } from '../../../types';
+
+export interface IToggleItem {
+  data: ServiceItem;
+  closeOverlay: CallableFunction;
+}
