+++ src/components/Sections/Navbar/Menu/MenuColumn/RouterNavItem/index.ts
@@ -1,0 +1,1 @@
+export { RouterNavItem } from './RouterNavItem';
