+++ src/components/Sections/Navbar/Menu/MenuColumn/RouterNavItem/types.ts
@@ -1,0 +1,6 @@
+import { RouterItem } from '../../types';
+
+export interface IRouterNavItem {
+  data: RouterItem;
+  closeOverlay: CallableFunction;
+}
