+++ src/components/Sections/Navbar/Menu/MenuColumn/index.ts
@@ -1,0 +1,1 @@
+export { MenuColumn } from './MenuColumn';
