+++ src/components/Sections/Navbar/Menu/MenuColumn/types.ts
@@ -1,0 +1,6 @@
+import { ColumnData } from '../types';
+
+export interface IColumnData {
+  data: ColumnData;
+  closeOverlay: CallableFunction;
+}
