+++ src/components/Sections/Navbar/Menu/index.ts
@@ -1,0 +1,2 @@
+export { Menu } from './Menu';
+export * from './types';
