--- src/components/Sections/Navbar/Overlay/index.ts
@@ -1,2 +1,0 @@
-export { Overlay } from './Overlay';
-export * from './types';
